requirements.txt
google-genai
google-search-results
python-dotenv
pandas
openpyxl
streamlit
serpapi